# src/main/resources/static/openapi.yaml
openapi: 3.1.0
info:
  title: Pathfinder 2e AI Oracle API
  description: |
    REST API for managing Foundry VTT PF2e data import and vector store ingestion.

    The system operates in two phases:
    1. **Import**: Download JSON data from the Foundry VTT PF2e GitHub repository
    2. **Ingestion**: Process imported data into the vector store for RAG-based similarity search

    Both operations run asynchronously and return job IDs for status tracking.
  version: 1.0.0
  contact:
    name: GitHub Repository
    url: https://github.com/larmic/Pathfinder-2e-AI-Oracle
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/import/jobs/{id}:
    get:
      tags:
        - Import
      summary: Get import job status
      description: Returns the current status of a specific import job.
      operationId: getImportJobStatus
      parameters:
        - name: id
          in: path
          required: true
          description: Job UUID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Job found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportJob'
        '404':
          description: Job not found
...
  /api/ingestion/all:
    post:
      tags:
        - Ingestion
      summary: Ingest entries into vector store
      description: |
        Triggers asynchronous ingestion of Foundry entries into the vector store for RAG-based similarity search.

        By default, performs incremental ingestion (only entries that have changed since last ingestion).
        Use `force=true` to perform a full re-index of all entries.

        **Note**: Full re-index may take 15-30 minutes depending on data size and embedding model.
      operationId: ingestAll
      parameters:
        - name: force
          in: query
          required: false
          description: If true, re-index all entries regardless of vectorization status
          schema:
            type: boolean
            default: false
      responses:
        '202':
          description: Ingestion job started
          headers:
            Location:
              description: URL to check job status
              schema:
                type: string
                example: /api/ingestion/jobs/019430b6-1234-7000-8000-000000000000
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionJob'
...
  /api/ingestion/jobs/{id}:
    get:
      tags:
        - Ingestion
      summary: Get ingestion job status
      description: Returns the current status of a specific ingestion job.
      operationId: getIngestionJobStatus
      parameters:
        - name: id
          in: path
          required: true
          description: Job UUID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Job found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionJob'
        '404':
          description: Job not found
...
components:
  schemas:
    ImportJob:
      type: object
      description: Represents an asynchronous import job
      properties:
        id:
          type: string
          format: uuid
          description: Unique job identifier
          example: "019430b6-1234-7000-8000-000000000000"
        itemType:
          type: string
          description: Type of import (always "ALL" for full import)
          example: "ALL"
        status:
          $ref: '#/components/schemas/JobStatus'
        createdAt:
          type: string
          format: date-time
          description: When the job was created
        startedAt:
          type: string
          format: date-time
          nullable: true
          description: When the job started processing
        completedAt:
          type: string
          format: date-time
          nullable: true
          description: When the job completed
        progress:
          $ref: '#/components/schemas/ImportProgress'
        result:
          $ref: '#/components/schemas/ImportResult'
        errorMessage:
          type: string
          nullable: true
          description: Error message if job failed
      required:
        - id
        - itemType
        - status
        - createdAt
        - progress
    ImportProgress:
      type: object
      description: Progress information for an import job
      properties:
        totalFiles:
          type: integer
          description: Total number of files to process
          example: 28000
        processedFiles:
          type: integer
          description: Number of files processed so far
          example: 14000
        skippedFiles:
          type: integer
          description: Number of files skipped (unchanged)
          example: 5000
        percentComplete:
          type: integer
          description: Completion percentage (0-100)
          example: 50
      required:
        - totalFiles
        - processedFiles
        - skippedFiles
        - percentComplete
    ImportResult:
      type: object
      nullable: true
      description: Final result of a completed import job
      properties:
        imported:
          type: integer
          description: Number of files imported
          example: 9000
        skipped:
          type: integer
          description: Number of files skipped
          example: 19000
        errors:
          type: integer
          description: Number of errors encountered
          example: 0
        durationSeconds:
          type: integer
          format: int64
          description: Total duration in seconds
          example: 300
        totalFiles:
          type: integer
          description: Total files processed
          example: 28000
      required:
        - imported
        - skipped
        - errors
        - durationSeconds
        - totalFiles
...
    IngestionJob:
      type: object
      description: Represents an asynchronous ingestion job
      properties:
        id:
          type: string
          format: uuid
          description: Unique job identifier
          example: "019430b6-1234-7000-8000-000000000000"
        foundryType:
          type: string
          description: Type of ingestion (always "ALL" for full ingestion)
          example: "ALL"
        status:
          $ref: '#/components/schemas/JobStatus'
        createdAt:
          type: string
          format: date-time
          description: When the job was created
        startedAt:
          type: string
          format: date-time
          nullable: true
          description: When the job started processing
        completedAt:
          type: string
          format: date-time
          nullable: true
          description: When the job completed
        progress:
          $ref: '#/components/schemas/IngestionProgress'
        result:
          $ref: '#/components/schemas/IngestionResult'
        errorMessage:
          type: string
          nullable: true
          description: Error message if job failed
      required:
        - id
        - foundryType
        - status
        - createdAt
        - progress
    IngestionProgress:
      type: object
      description: Progress information for an ingestion job
      properties:
        totalEntries:
          type: integer
          description: Total number of entries to process
          example: 28000
        processedEntries:
          type: integer
          description: Number of entries processed so far
          example: 14000
        errors:
          type: integer
          description: Number of errors encountered
          example: 5
        percentComplete:
          type: integer
          description: Completion percentage (0-100)
          example: 50
      required:
        - totalEntries
        - processedEntries
        - errors
        - percentComplete
    IngestionResult:
      type: object
      nullable: true
      description: Final result of a completed ingestion job
      properties:
        processed:
          type: integer
          description: Number of entries successfully processed
          example: 27995
        errors:
          type: integer
          description: Number of errors encountered
          example: 5
        total:
          type: integer
          description: Total entries attempted
          example: 28000
        durationSeconds:
          type: integer
          format: int64
          description: Total duration in seconds
          example: 1800
      required:
        - processed
        - errors
        - total
        - durationSeconds
...
    JobStatus:
      type: string
      description: Status of an asynchronous job
      enum:
        - PENDING
        - RUNNING
        - COMPLETED
        - FAILED
      example: "RUNNING"